PCI: epf-mhi: Add support for SA8775P
Add support for Qualcomm Snapdragon SA8775P SoC to the EPF driver.
SA8775P is currently reusing the PID 0x0306 (the default one hardcoded
in the config space header) as the unique PID is not yet allocated.
But the host side stack works fine with the default PID. It will get
updated once the PID is finalized. Currently, it has no fixed PCI class,
so it is being advertised as "PCI_CLASS_OTHERS".

Change-Id: I714357a7db0e2d810127c6372958caa00a7fd306
Signed-off-by: Mrinmay Sarkar <quic_msarkar@quicinc.com>
Patch-mainline: linux-pci @ 11/11/23, 08:03
Signed-off-by: Shashank Babu Chinta Venkata <quic_schintav@quicinc.com>