misc: dcc: Enable DCC dump in ramdump for mem map 3 devices
In the current state of the driver, DCC SRAM dump is triggered
in case of HW crashes only for memory map version 2 devices.

This patch adds the same support, to collect DCC register dump
from the full ram dump in case of a device crash, for devices
with memory map version 3. Verifed the same on SA8775p, which
is a memory map version 3 target.

Fixes: 042a41be9a53 ("FROMLIST: misc: dcc: Add driver support for Data Capture and Compare unit(DCC)")
Signed-off-by: Ninad Naik <quic_ninanaik@quicinc.com>
Change-Id: I1282221ce4113aa3b645d4994539a31e6085280e